Move config into YAML files
Load the config by merging together YAML files in the config directory instead
of having a bunch of constants lying around.

Pass the single config hash into the `Application` on initialization.

Also remove all default parameters from `Application` so that it's used in the
same way in test and production.

Conflicts:
	application.rb
	features/support/env.rb