Add ValidInputFilter and validators
These will replace `ProcessableInputFilter`. Naming the filter for the
fact it does input validation avoids giving the impression that anything
which makes it through `ProcessableInputFilter` is processable. Since
the `UniqueTagFilter` is applied afterwards and can call `unprocessable`
on the context, that's misleading.

Splitting the validation and the filter into separate objects allows us
to easily customise the validation for each endpoint by supplying a
different list of validators.