Keyword args cause symbol/string flip flop
All keys are now symbols as soon as the HTTP (Sinatra) boudary is crossed.

Previously keeping "data" as string symbols seemed better but both Sequel and
keyword args rely on / provide symbols.

This is no longer poses a memory problem as the params are whitelisted before
their keys are cast into symbols