Fix db:setup rake task
Connect to the `postgres` database instead of the `template1` database
to drop and create other databases. `template1` isn't really safe to
connect to, and if another user is connected to that database then the
command will fail.

Also stop doing this as the `postgres` user -- the default user on CI is
`jenkins`, and there's a Postgres user available for it.