Handle a topic already existing on GovDelivery's side.
Previously, when we attempted to create a subscriber list, if the topic already
exists with GovDelivery, the creation would fail "silently" and we'd get a bad
"SubscriberList" record stored in the email-alert-api.

Now, it will recognise that the list already exists, and populate the data
locally (using the list topics API) so that the error condition won't be
encountered again.

In addition, if *any* error condition is encountered with GovDelivery, an
exception will now be raised rather than it being swallowed silently.