Don't send bulletins if no topics match
If we try and send bulletins via govdelivery with an empty list of
topics, we get an error returned:

    GD-12004: To send a bulletin you must select at least one topic or
category that has subscribers

We'd like not to get that, so don't try sending the bulletin if there are
no matching topics.  Instead, add a log entry indicating that we didn't
try to send the bulletin.

Story: https://www.pivotaltracker.com/story/show/83203610