Don't attempt to fetch topic by name
Currently on staging, gov_delivery.read_topic_by_name takes 40 seconds
to return. This means that if a topic exists on govdelivery, but isn’t
in the email-alert-api database, then we attempt to fetch it, but that
will always result in an ambiguous timeout.

Until we can fix this, I think it is better to return an explicit
TopicAlreadyExistsError.