Rename policy SubscriberList links to 'parent'
The policy-publisher and govuk-content-schemas have been updated such
that email signup content items now carry a link to their parent content
item. The key used in the link is 'parent'. The implication of this is
that email-alert-api will start receiving links keyed with 'parent' for
policy email subscriptions and notifications. This commit supports that
transition.