Start reading from JSON fields.
Currently both hstore and json fields match, this
step switches the app over to reading from the
JSON fields.

Once this is shipped and happy, we'll switch the
hstore fields over to JSON fields with a full copy
of the data, switch to reading/writing those fields,
and then delete the *_json fields.

This should complete the migration with no downtime.