Fix duplicate subscriptions.
For *some reason* we have duplicate entries in the
subscriber list table for the same govdelivery
topic.

This removes all where the gov_delivery_id is null
which were specialist document subscriptions and
have been checked with Chris Patuzzo.

For the rest it removes all but the most recent.
The tags, links, and title all match for these, so
there should be no effect on the end user.

Finally, both Ruby and DB uniqueness constraints
are added to the subscriber list model.