Update GovDelivery topic IDs in integration
Despite integration and staging using separate accounts within GovDelivery’s staging environment, the ‘topic codes’ used for the topics must be unique across both environments.

If this constraint is not met when using the API to create topics and specifying the ID, the ID is silently ignored and new topic IDs are assigned to the topics. This results in either staging or integration acting very oddly depending on which tries to sync first.

The easiest way to fix this is to ensure we’re always using the correct account code as a prefix. This change updates all subscriber lists to change the prefix once the data has been synced from production.