Change handling of email alert titles
This commit splits the `title`, `short_name` and `description` attributes so that they can be sent separately to GovDelivery.

Since `title` and `short_name` have a 255 character limit enforced by GovDelivery, this commit adds the ability to pass a different, shorter string for these fields whilst keeping all the data in the `description` field.