Also delete topics which 404 as well as ones with 0 subscribers
Deleting these cleans up our data and also may fix some errors for users. If
we have a subscriber list in our database which doesn't exist in GovDelivery,
we'll redirect users to GovDelivery to sign up but that will fail because they
don't recognise the topic id. We don't have any of these right now in this
app, but since GovDelivery data can be modified in their UI it seems useful to
support this here as well as in govuk-delivery (where we have about 2000
topics in this state).

This changes the error-raising behaviour of the client but that shouldn't
cause any problems elsewhere in the application because TopicNotFound is a
subclass of UnknownError which should already be handled.

It's not very nice to have to check for a string zero, but now that we're
using parse_topic_response for the extra error handling we'd need to change
the behaviour of ResponseParser to fix this, and that's outside the scope of
this PR (but worth looking at separately).