Wrap notification logging in rescue block
The previous implementation could potentially have
resulted in subscribers receiving an email multiple
times if the NotificationLog create call failed. In
addition no NotificationLog would have been written
is send_bulletin raised an error.

By being a little more paranoid and wrapping the
logging code in a result block we can move it to
earlier in the processing stream thus ensuring that
we always log notification requests and that we never
send multiple notifications for a single change.