Add supertype handling
This enables the following functionality:

* create subscriber list with supertype criteria
* subscriber list filtering now honours supertype
criteria
* added testing around edge cases

In addition this change fixes a bug where searching
for a subscriber list with links or tags that failed
to find a match would fall back to searching by
document type (where links and tags are blank).

Example:
for a subscription type which relies on document type and
links/tags, such as travel advice for specific countries,
it wouldn't be possible to subscribe to travel advice for
a new country if there was already a list in the database
with only the document type. An API request to find a
subscriber list with

    links: {"countries": ["new-country-content-id"]}
    document_type: "travel_advice"

which doesn't exist in the database would instead return this
more general list if it exists (it's the all countries travel
advice list):

    document_type: "travel_advice"

This situation will happen only rarely for travel advice, but is much
more likely to arise when we have all the Whitehall subscriptions using
email-alert-api, so it's important to fix this.