Use `gov_delivery_id` when finding exact match
When we create a subscriber list via the api it first
does a lookup to see if a subscriber list with the
same criteria already exists.

With the migration of whitehall data it is possible
that multiple subscriber lists with different
gov_delivery_id will exists. As this is the case
it is important that when we are filtering we get
the correct record.

This is more important during the create process as
otherwise you would find a subscriber list with same
criteria, and would skip the creation of a record with
the desired gov_delivery_id resulting in users not
being notified.

example:

```
'https://www.gov.uk/government/feed?topics%5B%5D=uk-economy' has topic_id UKGOVUK_6888
'https://www.gov.uk/government/topics/uk-economy.atom' has topic_id UKGOVUK_8609
```