Support content items where the tags hash values are not arrays
This is required to support `specialist publisher` which send through
the `format` criteria in the `tags` hash as a string value insteade of
an array of strings.

We have revert this to the previous format rathger than fixing the
bug at source as we are not able to get 100% confidence that this
is the only source of data in this format.