Refactor TitleFetcher to fetch list of all topics
GovDelivery's API endpoint which lists all topics gives us all the data we
need to do this - we only need the topic code and name. As long as this one
request returns successfully, updating the data in the database should be very
quick, making the whole thing take much less time than if we fetch all topics
individually.

I'm leaving the previous version in the git history because the request to
fetch all the topics can take a long time (~30 seconds) and if it gets to the
point where it times out we may need revert back to the previous
implementation, which is more reliable but takes a lot longer.