Add documentation about subscriber list matching criteria
This gives an overview of how subscriber lists work. I haven't included many
examples because the tests are more comprehensive, but I've tried to explain
how the two sides of the app work and how the data matches up. The query code
can be complex to read so hopefully this will help people new to the app to
understand how it works.