Add a reporting script to help retire govuk-delivery
Before we can migrate, we need to ensure
email-alert-api behaves more-or-less the same as
govuk-delivery. This report provides detailed
output that we can use to fix problems with topic
matching between the two systems.

It also provides a ‘global’ view of which topics
are most frequently mismatched so we can focus our
attention on those.

https://trello.com/c/eKdM4dZr/44-improve-notification-reporting-when-comparing-email-alert-api-with-govuk-delivery-2