Fix `EmailSender::NotifySender` tests
There were a couple of issues with these tests that we missed on the
original PR that later caused a failure for some unknown reason and a
bit of confusion.

The original tests had some stubbing that masked issues and were also
accessing the subject through `Services.email_sender` which caused state
to bleed between tests due to the memoization in that module.

This commit removes the test for `false` being returned when an
exception is raised by the client as a) it is really hard to test as the
exception initializer needs a response passed into it and I couldn't
work out how to do it and b) we haven't actually implemented that
behaviour. The `send_email` method doesn't catch any exceptions so it
just raises or returns the client response.