Add email address validation
This just checks that the email address has an @ character in it,
which is easier than using a complex regular expression and it
won't reject any valid address.

Since we're going to be validating email addresses by sending
emails and asking the user to confirm they received the email
anyway, it's not important that this is too strict.