Add error handling to NotificationHandler
As we're currently running this alongside the 'live' production code we
should swallow all errors for now. This commit catches them and sends
them to sentry. It also adds a `version: 2` tag. We've not settled on
versioning the API as such yet but this will allow us to easily filter
errors on the code that is being actively developed.

We'll probably want to build on this and return error messages to the
caller as we go but this should be sufficient for now.