Add env var to specify email service provider
Allow selective use of either Notify or the Pseudo email service
provider based on a config value exposed as `email_service_provider`.

As a sensible starter, production will use Notify, other environments
will default to use the Pseudo service.

A new env var
EMAIL_SERVICE_PROVIDER is available which can be used to override the
default values, for example when working on the development vm, you
can force the use of Notify, though you will obviously need Notify
credentials set up as well.

Also set a limit to the size and number of log files that the Pseudo
email sender creates so we don't use too much space on the dev vm.