Send emails to courtesy subscribers
This allows us to have a list of email addresses who will always
receive a courtsey copy of every email which gets sent out.

This code is not meant to stay around for a long time, it should
get replaced by a more conrete solution rather than an array of
email addresses.