Add RateSleeper
The Notify API has a rate limit of 50 requests per second so we need to
throttle our delivery Sidekiq workers to respect this.

This class provides a simple way of achieving this. Within the workers
we will wrap the worker code with `RateSleeper.run` which measures the
duration of execution and then sleeps to maintain the appropriate rate.
The number requests we allow each worker to make per second will be set as an
environment variable using the formula:
 rate/s allowed by notify / number of workers (across all boxes)

e.g. 50 (requests/s) / 6 (workers) = 8 requests/s per worker

A default of 8 has been set.