Make each of the healthchecks influence overall status
Previously, the only healthchecks that influenced
the overall status of the /healtcheck endpoint
were the database and redis connections.

This change makes it so that the status is set to
the most severe status of the individual checks.
It also substantially refactors the controller and
extracts lots of Healthcheck models for each of
the checks.

Each of these models has a uniform interface
responding to #name, #status and #details and the
top-level Healthcheck also has the interface and
combines everything into one.