Only check the latest delivery attempt per email
If we re-deliver emails after a technical failure,
we don’t want the alert to remain in a
critical/warning state.

We could just ‘ack’ the alert, but then new
technical failures in this time period wouldn’t be
alerted on.

This change adds a new index which should help the
group operation when the delivery_attempts table
is large. We should keep an eye on this query time
As the database grows.