Mark email processed
This commit adds a call to `Email#mark_processed!` in the
`DeliverEmail` command. We did consider doing this within a transaction
but having thought about it a partially complete set of data and an
exception reported would be more useful than just the exception so I've
omitted the transaction for now.