Make sure priority is passed as a symbol
Currently, the priority gets turned into a string by Sidekiq so we
need to turn it back into a symbol when using it. I spotted this
issue by deploying the app in staging.

There is actually a test that would have caught this, but we send
all exceptions to Sentry rather than raising them. This still
feels right to have for the moment, so I decided against adding
a new test for this that would then just get removed when we
remove explicit Sentry integration.