Add an Unsubscribe service that encapsulates unsubscribe behaviour
There are two use-cases for unsubscribing:

1) When we receive a permanent-failure from Notify
2) When the user clicks ‘unsubscribe’ from an email

The former works at a subscriber level whereas the
latter works on a subscription. There’s some
dependency between these actions because we need
to nullify the subscriber’s email address if it’s
the last subscription.

Therefore, we thought it made sense to pull this
code together and remove the responsibilities from
the models themselves.