Pass subscriber instead of address
We were passing the email address of a subscriber into the `Email`
initializer. To implement unsubscribe links we need to be able to get a
list of the subscriptions as well so this commit passes the `Subscriber`
object instead and uses that to retrieve the address.

We will then be able to use the `Subscriber` object to retrieve all of
the users subscriptions too.