Store Email timing statistics
This creates timing statistics with statsd for when emails are first
sent to a third party (notify). They both measure the time difference
between the object created (Email, ContentChange) and what time it is
that they are first sent. This is to allow us to have data on how long
it takes for things to enter our system and then be sent to the next
system.

It uses the hostname in the key sent to statsd to work around this
issue: https://github.com/alphagov/govuk-puppet/pull/6834

This has a somewhat crude condition to exclude digests exist as they will skew
the data. Hopefully at that point we will have a clearer means to
calculate them.