Prefix subject for non-production environments
We need to be able to send courtesy copies and some subscribed emails in
the integration and staging environments to allow us to test changes
before deployment. This commit introduces a prefix (INTEGRATION |
STAGING) to the email subject so as we can tell the difference in the
courtesy copy account.