Add subscribables#show endpoint
We need to be able to retrieve a `SubscriberList` by gov_delivery_id to
allow us to retrieve it from the frontend.

There is an existing `SubscriberListsController#show` but when only the
gov_delivery_id param is passed it always returns the same
`SubscriberList`. This could be considered a bug but we will be changing
`SubscriberList` to `Subscribable` at some point soon and to avoid
changing existing code that is running in production until we're ready
to replace this I've added this `Subscribables` controller.

`SubscriberList` serialises to `subscriber_list: <attributes>` so there
is a bit of funkiness to avoid that.