Remove duplicate SubscriberLists
There are a few `SubscriberList` records that are duplicated. This has
come to light as we move towards using what is currently called
`gov_delivery_id` as a unique id.

This commit removes the duplicates leaving only one record for each
`gov_delivery_id` and adds a unique constraint on `gov_delivery_id`.