Remove duplicate gov_delivery_id use
We've added a unique DB constraint which is causing these tests to
fail.

This commit removes the explicit `gov_delivery_id` duplicates and lets
the factory create unique ones. I don't see why they needed to be
duplicate in the first place and the tests still pass but they were
explicitly marked out as `DUPLICATE` as if that was important.