Remove redundant test set up
We have added a unique constraint on `gov_delivery_id` which has caused
one of the tests in `data_sync_spec.rb` to fail. This test was trying to create
two `SubscriberList` objects with the same `gov_delivery_id` but one of
the objects appears redundant as it isn't returned by the
gov_delivery_client in the test set up so doesn't affect the test.

This commit deletes it, the test now passes.