Add a feature flag to use email-alert-frontend for email collection
https://trello.com/c/bwRGOWZQ/409-add-feature-to-apps-that-serve-email-signup-pages-to-redirect-to-email-alert-frontend-for-email-address-collection

We serve the subscription_url as a field in the
JSON response when clients call
find_or_create_subscriber_list. That means we can
toggle the subscription_url in email-alert-api to
affect all apps.

If we need more granular control, we can switch
over a few document_types at a time.

Here's where we use `subscription_url` in each client app:

- https://github.com/alphagov/collections/blob/60fcda67508170fe51b8227139fe217bfa47ca65/app/models/email_signup.rb#L16
- https://github.com/alphagov/finder-frontend/blob/b0fd3ce2aa6405dc9ffd50912b832e471cc7cb37/lib/email_alert_signup_api.rb#L14
- https://github.com/alphagov/whitehall/blob/a53f622a3e8fb5aa296bc532ae0d48a29e614a38/app/models/email_signup.rb#L41