Configure delivery workers to use a separate queue
This will mean that 10 other workers are picked for every 5
high priority delivery worker and every 1 low priority delivery
worker.

The aim of this is to solve a problem we've noticed in load
testing where the delivery request workers are sleeping a lot
which prevents other workers from performing actions.