Reverse prioritisation of workers
Currently we are prioritising the other workers over the delivery
workers to avoid the delivery workers from sleeping and wasting
time. After doing some smaller experiments, what I've learnt is
that to achieve a much better throughput of emails with the other
workers, we don't actually need to prioritise the other workers
as I have been doing, we just need to leave a few spaces free for
the other workers while still prioritising the delivery request
workers.