Add a test to check for duplicate emails being sent out
We can't test that the correct number of emails are being sent out
as the lock is exclusive across processes, not threads. However,
we can test that when the lock doesn't work, we get an exception
from the optimistic locking meaning that duplicate emails will
still not be sent out and the worker will be retried.