Set a fake Notify api_key / base_url in dev/test
If we set these credentials to nil, we can't
instantiate the Notify client in the test-suite
which is problematic when setting message
expectations on it.

We should leave production alone as it’s probably
a good thing not to be able to instantiate the
client if there are no credentials.