Allow nullable: subscription_contents.subscription_id
We delete subscription records when users
unsubscribe. This means we need to make this field
nullable on the subscription_contents table.

We considered also deleting the
subscription_contents associated with the deleted
subscription, but decided not to because the
subscription_content still points to the
content_change which is useful information.