Send ‘reference’ and refactor services into DeliveryRequestService
This supersedes:
- DeliverEmailService
- EmailSenderService

The name of this service matches the name of the
worker: DeliveryRequestWorker

The ‘reference’ is just a string. I tried to
encode some helpful information in it:
- the email id
- a timestamp

Rather than having to use mocks everywhere, I’ve
made it so the config is injected, in the same
way as the NotifyProvider.

I’ve tried to extract EmailAddressOverrider into
a separate class to hide some of these details
from the high-level flow. We could put this in a
separate file if we want.