Treat requests as JSON
This sets all our requests to be treated as a JSON format, which means
we get JSON exceptions.

The config.debug_exception_response_format sets the debug exceptions to
be rendered in JSON rather than HTML.

This would all be slightly neater if we used Rails api_only mode however
GDS SSO is currently incompatible.