Scope routes to default to format json without others
This defaults all of our defined routes to be a format of JSON without
the user having to specify it - which makes sense since we don't intend
to return non JSON responses.

It also disables the ability to specify a format so we now only serve
routes such as /healthcheck rather than having the ability to have
/healthcheck. /healthcheck.json, /healthcheck.xml etc. Which seems
superfluous for this application.