Add a unique index on subscriber list titles
GovDelivery currently requires these to be unique anyway and it
seems unlikely that we will want subscriber lists with the same
name.

The lack of a unique constraint here was causing a problem with
the data sync task as it relies on unique titles and GovDelivery
IDs to do a difference on arrays to decide which subscriber lists
to delete.