Fix an autoload problem for symbolize_json
On my development machine, when running the
email-alert-api server, I was seeing the
healthcheck integration test fail.

The failing spec was caused by a non-zero Sidekiq
retry set size. The failing Sidekiq jobs indicated
a problem loading the symbolize_json file.

I’m not sure why this only happens when the server
is running. It could be to do with Spring, but
moving this file to a directory on the
eager_load_path fixes it.