Add DailyDigestSchedulerService
Adds a basic implementation of a service that creates a DigestRun.

The creation of the `DigestRun` is expected to happen only once
for a given period regardless of how many times this service is called
so it is protected in this commit by an advisory lock and a presence
check.

This will be ultimately be called on a schedule.