Store signon user id when actions are performed
For auditing purposes this will store the user id of the user that has
performed an action which has altered the state of the application.

Since we have authentication we have the ability to know who has
performed an action, but this was previously not stored. Storing this
should help if we want to work out analytics based on a user service.

For the most part though this data should be pretty dull with mostly the
same accounts using the API.