Return distinct subscribers from query
The `DigestRunSubscriberQuery` returns all subscribers that have
matching content within a the digest run range. This query was returning
the same subscriber multiple times if they were subscribed to more than
one `Subscribable` that matched content. This was causing the subscriber
to receive multiple digests instead of just one.

This commit adds a `.distinct` to the query which removes the
duplicates.