Soft delete `Subscription`
While we are migrating we want to maintain a record of deleted
subscriptions (unsubscribes) to allow us to rollback and restore
previous state if need be. This will be reverted post migration to
comply with our data retention policy.

This commit adds a `deleted_at` field and updates all calling code to
use a new `not_deleted` scope when retrieving `Subscription` records.