Set up environment variable for a whitelisting email addresses
This environment variable would only be of any use when we have a value
set for `ENV["EMAIL_ADDRESS_OVERRIDE"]`. It would be used to allow one
or more email addresses specified by the EMAIL_ADDRESS_OVERRIDE_WHITELIST value
to send emails to actual users.

This variable would be a comma separated list of email-addresses e.g
```
EMAIL_ADDRESS_OVERRIDE_WHITELIST=kevin.dew@example.com,joe.bloggs@example.com
```

The intention for this feature is for when there are cases when by and
large you don't want any emails actually sent to users but you do want
to have a select group of test users who can receive an emails.

An example would be in the time leading up to the launch of the switch
to Notify. In that scenario we don't want any subscribed users to
receive emails but we do want a handful of internal people to get the
emails for testing.

This could also be useful for staging and integration environments.