Some refactoring of DeliveryRequestService
This class seemed somewhat inconsistent with other classes in this
application with it's usage of attr_accessor rather than attr_reader -
so that has been switched around.

The EmailAddressOverrider class has been simplified to not have the
logging in (which has been moved to DeliveryRequestService) this
simplifies the concerns of the overrider in preparation for the
overrider whitelist.

The log has also been simplified as it seemed a bit strange to be
outputting almost the exact same log for this and then again in the
psuedo class (if that's used).