Remove load testing code
We're no longer doing any load testing so it seems fine to remove
this code. It quickly becomes out of date as the rest of the code
changes, so if a user wanted to run a load test again they would
have to make some code changes, and reverting this PR would only be
a small part of that.