Remove fake data code
We're now using the GovDelivery import script to generate fake
data so it seems fine to remove this. It can quickly end up out
of date as a rarely used part of the system, so if a user wanted
to run this again they're likely to have to make code changes, and
reverting this PR would only be a small part of that.