Ignore duplication notifications
In the event that `email-alert-service` or one of the other client apps
has a request timeout and retries we don't want to process the request
multiple times once we've received it.

We currently go on to handle a notification via two paths, the
govdelivery one and the notify one. Once we just do this via Notify we
should probably handle this as a validation on `ContentChange` and a
unique constraint on a composite key in the DB. We will need to clean up
the data for that as there are already dupes in there.